@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: ./api/utils/auth/auth_provider.go
+//
+// Generated by this command:
+//
+//	mockgen -source ./api/utils/auth/auth_provider.go -destination ./api/test/mock/auth_provider_mock.go -package mock
+//
 
 // Package mock is a generated GoMock package.
 package mock
@@ -9,13 +14,14 @@ import (
 	reflect "reflect"
 
 	auth "github.com/equinor/radix-cost-allocation-api/api/utils/auth"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockAuthProvider is a mock of AuthProvider interface.
 type MockAuthProvider struct {
 	ctrl     *gomock.Controller
 	recorder *MockAuthProviderMockRecorder
+	isgomock struct{}
 }
 
 // MockAuthProviderMockRecorder is the mock recorder for MockAuthProvider.
@@ -45,7 +51,7 @@ func (m *MockAuthProvider) VerifyToken(ctx context.Context, token string) (auth.
 }
 
 // VerifyToken indicates an expected call of VerifyToken.
-func (mr *MockAuthProviderMockRecorder) VerifyToken(ctx, token interface{}) *gomock.Call {
+func (mr *MockAuthProviderMockRecorder) VerifyToken(ctx, token any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "VerifyToken", reflect.TypeOf((*MockAuthProvider)(nil).VerifyToken), ctx, token)
 }
@@ -54,6 +60,7 @@ func (mr *MockAuthProviderMockRecorder) VerifyToken(ctx, token interface{}) *gom
 type MockIDToken struct {
 	ctrl     *gomock.Controller
 	recorder *MockIDTokenMockRecorder
+	isgomock struct{}
 }
 
 // MockIDTokenMockRecorder is the mock recorder for MockIDToken.
@@ -74,7 +81,7 @@ func (m *MockIDToken) EXPECT() *MockIDTokenMockRecorder {
 }
 
 // GetClaims mocks base method.
-func (m *MockIDToken) GetClaims(out interface{}) error {
+func (m *MockIDToken) GetClaims(out any) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetClaims", out)
 	ret0, _ := ret[0].(error)
@@ -82,7 +89,7 @@ func (m *MockIDToken) GetClaims(out interface{}) error {
 }
 
 // GetClaims indicates an expected call of GetClaims.
-func (mr *MockIDTokenMockRecorder) GetClaims(out interface{}) *gomock.Call {
+func (mr *MockIDTokenMockRecorder) GetClaims(out any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetClaims", reflect.TypeOf((*MockIDToken)(nil).GetClaims), out)
 }
