@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: ./models/radix_api/client.go
+//
+// Generated by this command:
+//
+//	mockgen -source ./models/radix_api/client.go -destination ./api/test/mock/radix_api_client_mock.go -package mock
+//
 
 // Package mock is a generated GoMock package.
 package mock
@@ -10,13 +15,14 @@ import (
 	radix_api "github.com/equinor/radix-cost-allocation-api/models/radix_api"
 	application "github.com/equinor/radix-cost-allocation-api/models/radix_api/generated_client/client/application"
 	platform "github.com/equinor/radix-cost-allocation-api/models/radix_api/generated_client/client/platform"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockRadixAPIClient is a mock of RadixAPIClient interface.
 type MockRadixAPIClient struct {
 	ctrl     *gomock.Controller
 	recorder *MockRadixAPIClientMockRecorder
+	isgomock struct{}
 }
 
 // MockRadixAPIClientMockRecorder is the mock recorder for MockRadixAPIClient.
@@ -46,7 +52,7 @@ func (m *MockRadixAPIClient) GetRadixApplicationDetails(appParams *application.G
 }
 
 // GetRadixApplicationDetails indicates an expected call of GetRadixApplicationDetails.
-func (mr *MockRadixAPIClientMockRecorder) GetRadixApplicationDetails(appParams, token interface{}) *gomock.Call {
+func (mr *MockRadixAPIClientMockRecorder) GetRadixApplicationDetails(appParams, token any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetRadixApplicationDetails", reflect.TypeOf((*MockRadixAPIClient)(nil).GetRadixApplicationDetails), appParams, token)
 }
@@ -61,7 +67,7 @@ func (m *MockRadixAPIClient) ShowRadixApplications(appParams *platform.ShowAppli
 }
 
 // ShowRadixApplications indicates an expected call of ShowRadixApplications.
-func (mr *MockRadixAPIClientMockRecorder) ShowRadixApplications(appParams, token interface{}) *gomock.Call {
+func (mr *MockRadixAPIClientMockRecorder) ShowRadixApplications(appParams, token any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShowRadixApplications", reflect.TypeOf((*MockRadixAPIClient)(nil).ShowRadixApplications), appParams, token)
 }
