@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: ./repository/repository.go
+//
+// Generated by this command:
+//
+//	mockgen -source ./repository/repository.go -destination ./repository/mock/repository_mock.go -package mock
+//
 
 // Package mock is a generated GoMock package.
 package mock
@@ -9,13 +14,14 @@ import (
 	time "time"
 
 	models "github.com/equinor/radix-cost-allocation-api/models"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockRepository is a mock of Repository interface.
 type MockRepository struct {
 	ctrl     *gomock.Controller
 	recorder *MockRepositoryMockRecorder
+	isgomock struct{}
 }
 
 // MockRepositoryMockRecorder is the mock recorder for MockRepository.
@@ -45,7 +51,7 @@ func (m *MockRepository) GetContainers(from, to time.Time) ([]models.ContainerDt
 }
 
 // GetContainers indicates an expected call of GetContainers.
-func (mr *MockRepositoryMockRecorder) GetContainers(from, to interface{}) *gomock.Call {
+func (mr *MockRepositoryMockRecorder) GetContainers(from, to any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetContainers", reflect.TypeOf((*MockRepository)(nil).GetContainers), from, to)
 }
