@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: ./service/costservice.go
+//
+// Generated by this command:
+//
+//	mockgen -source ./service/costservice.go -destination ./service/mock/costservice.go -package mock
+//
 
 // Package mock is a generated GoMock package.
 package mock
@@ -9,13 +14,14 @@ import (
 	time "time"
 
 	models "github.com/equinor/radix-cost-allocation-api/models"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockCostService is a mock of CostService interface.
 type MockCostService struct {
 	ctrl     *gomock.Controller
 	recorder *MockCostServiceMockRecorder
+	isgomock struct{}
 }
 
 // MockCostServiceMockRecorder is the mock recorder for MockCostService.
@@ -45,7 +51,7 @@ func (m *MockCostService) GetCostForPeriod(from, to time.Time) (*models.Applicat
 }
 
 // GetCostForPeriod indicates an expected call of GetCostForPeriod.
-func (mr *MockCostServiceMockRecorder) GetCostForPeriod(from, to interface{}) *gomock.Call {
+func (mr *MockCostServiceMockRecorder) GetCostForPeriod(from, to any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCostForPeriod", reflect.TypeOf((*MockCostService)(nil).GetCostForPeriod), from, to)
 }
@@ -60,7 +66,7 @@ func (m *MockCostService) GetFutureCost(appName string) (*models.ApplicationCost
 }
 
 // GetFutureCost indicates an expected call of GetFutureCost.
-func (mr *MockCostServiceMockRecorder) GetFutureCost(appName interface{}) *gomock.Call {
+func (mr *MockCostServiceMockRecorder) GetFutureCost(appName any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetFutureCost", reflect.TypeOf((*MockCostService)(nil).GetFutureCost), appName)
 }
