@@ -2,25 +2,25 @@
 //
 // This is the Cost allocation API Server for the Radix platform.
 //
-//     Schemes: http, https
-//     BasePath: /api/v1
-//     Version: 1.1.1
-//     Contact: https://equinor.slack.com/messages/CBKM6N2JY
+//	Schemes: http, https
+//	BasePath: /api/v1
+//	Version: 1.1.1
+//	Contact: https://equinor.slack.com/messages/CBKM6N2JY
 //
-//     Consumes:
-//     - application/json
+//	Consumes:
+//	- application/json
 //
-//     Produces:
-//     - application/json
+//	Produces:
+//	- application/json
 //
-//     Security:
-//     - bearer:
+//	Security:
+//	- bearer:
 //
-//     SecurityDefinitions:
-//     bearer:
-//          type: apiKey
-//          name: Authorization
-//          in: header
+//	SecurityDefinitions:
+//	bearer:
+//	     type: apiKey
+//	     name: Authorization
+//	     in: header
 //
 // swagger:meta
 package docs
